Fagprojekt_DayTrading/pyproject.toml
[build-system]
requires = ["uv_build>=0.9.22,<0.10.0"]
build-backend = "uv_build"

[project]
name = "kvant"
version = "0.0.1"
description = "Research project on automated day trading using short-term (1\u20135 min) stock data. Develops and evaluates a risk-aware, online trading algorithm that jointly optimizes filtering and trade-entry decisions for profit under transaction costs. Implemented and backtested in Python with vectorbt under realistic, non-stationary market conditions."
authors = [
  { name = "group 1", email = "your@email.com" },
]
license = { file = "LICENSE" }
keywords = ["machine learning", "MLOps"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Programming Language :: Python :: 3",
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=4.6.0",
    "fastapi==0.115.6",
    "huggingface-hub>=1.4.1",
    "pandas>=3.0.1",
    "pyarrow>=23.0.1",
    "python-dotenv>=1.2.2",
    "torch==2.6.0",
    "typer==0.15.1",
    "uvicorn==0.34.0",
    "vectorbt>=0.28.2",
]

[dependency-groups]
dev = [
    "coverage==7.6.9",
    "invoke==2.2.0",
    "pre-commit==4.1.0",
    "pytest==8.3.4",
    "ruff==0.1.3",
    "mypy==1.19.1",
    "mkdocs-material==9.7.1",
    "mkdocstrings-python==2.0.1",
]


[tool.uv.sources]
torch = [
    { index = "pytorch-cpu" },
]
torchvision = [
    { index = "pytorch-cpu" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.ruff]
line-length = 120

[tool.coverage.run]
omit = ["tests/*"]
